refactor : job와 step 기능별 클래스로 추출
